Expanded Crouzeix-Raviart reffe names like Raviart-Thomas
and cleant associated test, removed a dev file that doesn't belong to
the repo
